--- a/sheldon/bin/setup.sh
+++ b/sheldon/bin/setup.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+mkdir -p "$HOME/.config/sheldon"
+ln -s "$PWD/sheldon/config.d/plugins.toml" "$HOME/.config/sheldon/plugins.toml"