```zsh
> git checkout main
Switched to branch 'main'
Your branch is up to date with 'origin/main'.

> for i in `seq 5`; do; time zsh -i -c exit; done
zsh -i -c exit  1.08s user 0.44s system 53% cpu 2.829 total
zsh -i -c exit  1.09s user 0.44s system 53% cpu 2.851 total
zsh -i -c exit  1.09s user 0.44s system 53% cpu 2.871 total
zsh -i -c exit  1.09s user 0.45s system 53% cpu 2.901 total
zsh -i -c exit  1.09s user 0.44s system 53% cpu 2.888 total
```

```zsh
> git checkout performance-tuning
Switched to branch 'performance-tuning'

> for i in `seq 5`; do; time zsh -i -c exit; done
Loaded ~/.config/sheldon/plugins.toml
   Checked https://github.com/zsh-users/zsh-autosuggestions
   Checked https://github.com/zsh-users/zsh-syntax-highlighting
   Checked https://github.com/rupa/z
Locked ~/.local/share/sheldon/plugins.lock
zsh -i -c exit  0.14s user 0.21s system 59% cpu 0.588 total
zsh -i -c exit  0.13s user 0.19s system 74% cpu 0.423 total
zsh -i -c exit  0.13s user 0.19s system 74% cpu 0.424 total
zsh -i -c exit  0.13s user 0.19s system 70% cpu 0.446 total
zsh -i -c exit  0.13s user 0.19s system 75% cpu 0.419 total
```